This fixes a problem when the proxy is run the docker container and fails to get the defaults from kingpin
